# template/api/system/v1/file_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v6.31.1
// source: system/v1/file.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationFileFileUploadCallback = "/api.system.v1.File/FileUploadCallback"
const OperationFileGenFileUploadUrl = "/api.system.v1.File/GenFileUploadUrl"

type FileHTTPServer interface {
	// FileUploadCallback 文件上传完成回调接口
	FileUploadCallback(context.Context, *FileUploadCallbackRequest) (*FileUploadCallbackReply, error)
	// GenFileUploadUrl 生成文件上传URL
	GenFileUploadUrl(context.Context, *GenFileUploadUrlRequest) (*GenFileUploadUrlReply, error)
}

func RegisterFileHTTPServer(s *http.Server, srv FileHTTPServer) {
	r := s.Route("/")
	r.POST("/api/v1/file/upload-url", _File_GenFileUploadUrl0_HTTP_Handler(srv))
	r.POST("/api/v1/file/upload/finished", _File_FileUploadCallback0_HTTP_Handler(srv))
}

func _File_GenFileUploadUrl0_HTTP_Handler(srv FileHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GenFileUploadUrlRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFileGenFileUploadUrl)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GenFileUploadUrl(ctx, req.(*GenFileUploadUrlRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GenFileUploadUrlReply)
		return ctx.Result(200, reply)
	}
}

func _File_FileUploadCallback0_HTTP_Handler(srv FileHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in FileUploadCallbackRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFileFileUploadCallback)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.FileUploadCallback(ctx, req.(*FileUploadCallbackRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*FileUploadCallbackReply)
		return ctx.Result(200, reply)
	}
}

type FileHTTPClient interface {
	// FileUploadCallback 文件上传完成回调接口
	FileUploadCallback(ctx context.Context, req *FileUploadCallbackRequest, opts ...http.CallOption) (rsp *FileUploadCallbackReply, err error)
	// GenFileUploadUrl 生成文件上传URL
	GenFileUploadUrl(ctx context.Context, req *GenFileUploadUrlRequest, opts ...http.CallOption) (rsp *GenFileUploadUrlReply, err error)
}

type FileHTTPClientImpl struct {
	cc *http.Client
}

func NewFileHTTPClient(client *http.Client) FileHTTPClient {
	return &FileHTTPClientImpl{client}
}

// FileUploadCallback 文件上传完成回调接口
func (c *FileHTTPClientImpl) FileUploadCallback(ctx context.Context, in *FileUploadCallbackRequest, opts ...http.CallOption) (*FileUploadCallbackReply, error) {
	var out FileUploadCallbackReply
	pattern := "/api/v1/file/upload/finished"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationFileFileUploadCallback))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GenFileUploadUrl 生成文件上传URL
func (c *FileHTTPClientImpl) GenFileUploadUrl(ctx context.Context, in *GenFileUploadUrlRequest, opts ...http.CallOption) (*GenFileUploadUrlReply, error) {
	var out GenFileUploadUrlReply
	pattern := "/api/v1/file/upload-url"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationFileGenFileUploadUrl))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}