# template/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /api/v1/file/upload-url:
        post:
            tags:
                - File
            description: 生成文件上传URL
            operationId: File_GenFileUploadUrl
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.GenFileUploadUrlRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.GenFileUploadUrlReply'
    /api/v1/file/upload/finished:
        post:
            tags:
                - File
            description: 文件上传完成回调接口
            operationId: File_FileUploadCallback
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.FileUploadCallbackRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.FileUploadCallbackReply'
...
    /api/v1/system/auth/login:
        post:
            tags:
                - Auth
            description: 用户登录
            operationId: Auth_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.LoginReply'
    /api/v1/system/auth/logout:
        post:
            tags:
                - Auth
            description: 用户登出
            operationId: Auth_Logout
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.LogoutRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.LogoutReply'
...
components:
    schemas:
...
        api.system.v1.FileUploadCallbackReply:
            type: object
            properties:
                code:
                    type: integer
                    description: 处理状态码，0表示成功
                    format: int32
                message:
                    type: string
                    description: 处理状态信息
                fileUrl:
                    type: string
                    description: 文件访问URL
            description: 文件上传完成回调响应
        api.system.v1.FileUploadCallbackRequest:
            type: object
            properties:
                bucket:
                    type: string
                    description: 存储桶名称
                object:
                    type: string
                    description: 文件对象名称(key)
                etag:
                    type: string
                    description: 文件ETag
                size:
                    type: string
                    description: 文件大小(字节)
                mimeType:
                    type: string
                    description: 文件MIME类型
                imageInfo:
                    allOf:
                        - $ref: '#/components/schemas/api.system.v1.FileUploadCallbackRequest_ImageInfo'
                    description: 图片信息
                customVars:
                    type: object
                    additionalProperties:
                        type: string
                    description: 自定义变量
            description: 文件上传完成回调请求
        api.system.v1.FileUploadCallbackRequest_ImageInfo:
            type: object
            properties:
                height:
                    type: integer
                    format: int32
                width:
                    type: integer
                    format: int32
                format:
                    type: string
            description: 图片信息(如果是图片)
...
        api.system.v1.GenFileUploadUrlReply:
            type: object
            properties:
                sts:
                    $ref: '#/components/schemas/api.system.v1.GenFileUploadUrlReply_Sts'
                url:
                    $ref: '#/components/schemas/api.system.v1.GenFileUploadUrlReply_Url'
            description: 生成文件上传URL的响应
        api.system.v1.GenFileUploadUrlReply_Sts:
            type: object
            properties:
                accessKeyId:
                    type: string
                accessKeySecret:
                    type: string
                securityToken:
                    type: string
                region:
                    type: string
                bucketName:
                    type: string
                expiration:
                    type: string
                objectName:
                    type: string
                endpoint:
                    type: string
        api.system.v1.GenFileUploadUrlReply_Url:
            type: object
            properties:
                url:
                    type: string
                method:
                    type: string
                expiration:
                    type: string
                signedHeaders:
                    type: object
                    additionalProperties:
                        type: string
        api.system.v1.GenFileUploadUrlRequest:
            type: object
            properties:
                fileName:
                    type: string
                    description: 文件名
                fileSize:
                    type: string
                    description: 文件大小（字节数）
                mimeType:
                    type: string
                    description: MIME类型(如image/jpeg)
                bizType:
                    type: integer
                    description: 桶权限类型(1=公开桶/2=私有桶)
                    format: int32
                usageType:
                    type: string
                    description: 业务用途(如avatar/contract/product_img,可选)
                forceSts:
                    type: boolean
                    description: 强制使用STS方式上传
                callback:
                    allOf:
                        - $ref: '#/components/schemas/api.system.v1.GenFileUploadUrlRequest_Callback'
                    description: 回调配置，如果设置会覆盖全局配置
            description: 生成文件上传URL的请求
        api.system.v1.GenFileUploadUrlRequest_Callback:
            type: object
            properties:
                enable:
                    type: boolean
                    description: 是否启用回调
                url:
                    type: string
                    description: 回调URL
                body:
                    type: string
                    description: |-
                        回调请求体，支持变量替换: $(key), $(bucket), $(etag), $(size),
                         $(mimeType), $(imageInfo)等
            description: 回调相关参数
...
        api.system.v1.LoginReply:
            type: object
            properties:
                token:
                    type: string
                    description: 访问令牌
                expires:
                    type: string
                    description: 过期时间（秒）
            description: 登录响应
        api.system.v1.LoginRequest:
            type: object
            properties:
                username:
                    type: string
                    description: 用户名
                password:
                    type: string
                    description: 密码
            description: 登录请求
        api.system.v1.LogoutReply:
            type: object
            properties: {}
            description: 登出响应
        api.system.v1.LogoutRequest:
            type: object
            properties: {}
            description: 登出请求
...